update rounding test

+++ tests/test_rounding.cpp
@@ -21,7 +21,7 @@ using std::to_string;
 
 extern "C" {
 	void decompose_jazz(int32_t a, int32_t* a0, int32_t* a1);
-	uint8_t make_hint_jazz(int32_t a0, int32_t a1);
+	uint32_t make_hint_jazz(int32_t a0, int32_t a1);
 }
 
 uint8_t sample() {
@@ -54,7 +54,7 @@ int main() {
 		}
 
 		unsigned int hint_ref = make_hint(a0_ref, a1_ref);
-		uint8_t hint_jazz = make_hint_jazz(a0_jazz, a1_jazz);
+		uint32_t hint_jazz = make_hint_jazz(a0_jazz, a1_jazz);
 
 		if(hint_ref != hint_jazz) {
 			PRINT(a0_ref);
